refactor: move dataTypes to Constants
