+++ src/Constants/index.Constants.js
@@ -1,7 +1,8 @@
-import { TOTAL_CONFIRMED } from '../Observer/dataTypes';
+import { TOTAL_CONFIRMED } from './dataTypes';
 
 export const CORS_ANYWHERE = 'https://cors-anywhere.herokuapp.com/';
 export const COVID_API = 'https://api.covid19api.com';
 export const COUNTRY_INFO_API = 'https://restcountries.eu/rest/v2';
 export const DEFAULT_DATA_TYPE = TOTAL_CONFIRMED;
 export const DEFAULT_COUNTRY = 'ALL';
+export const DEFAULT_PRECISION = 2;
